requirements.txt
Pillow==2.8.1
wheel==0.24.0
